exabot-backend/requirements.txt
# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20

# Database
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Redis
redis==5.2.1

# Task Queue
celery==5.4.0
celery[redis]==5.4.0

# Auth & Security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==44.0.0

# HTTP Client
httpx==0.28.1

# Validation & Settings
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1

# AI/ML
openai==1.58.1
transformers==4.47.1
torch==2.5.1
scikit-learn==1.6.0
xgboost==2.1.3
numpy==2.2.0
pandas==2.2.3

# File Storage (S3-compatible)
boto3==1.35.93

# Monitoring & Logging
structlog==24.4.0

# Utilities
email-validator==2.2.0
python-dateutil==2.9.0
pytz==2024.2
